Fix a big in Obfuscate with 64-bit longs
Or maybe this is more like enable a bug because signed shifts cause one of the values to become 0xFFFFFFFF in the middle

+++ CODE/init.cpp
@@ -2265,13 +2265,13 @@ long Obfuscate(char const * string)
 	**	Transform the buffer into a number. This transformation is character
 	**	order dependant.
 	*/
-	long code = Calculate_CRC(buffer, length);
+	int32_t code = Calculate_CRC(buffer, length);
 
 	/*
 	**	Record a copy of this initial transformation to be used in a later
 	**	self referential transformation.
 	*/
-	long copy = code;
+	int32_t copy = code;
 
 	/*
 	**	Reverse the character string and combine with the previous transformation.
@@ -2368,7 +2368,7 @@ long Obfuscate(char const * string)
 	/*
 	**	Return the final code value.
 	*/
-	return(code);
+	return((uint32_t)code);
 }
